refactor: rename geometry.py to diffractometer.py (#213)
Rename the module so the filename matches the AdHocDiffractometer class
it contains.  Update all imports, Sphinx cross-references, docstrings,
documentation, CI workflow paths, AGENTS.md, roadmap.md, and test files.

Contributed by: OpenCode (argo/claudeopus46)
