+++ .github/workflows/benchmark.yml
@@ -12,7 +12,7 @@ on:
       - "src/ad_hoc_diffractometer/reference.py"
       - "src/ad_hoc_diffractometer/presets.py"
       - "src/ad_hoc_diffractometer/factories.py"
-      - "src/ad_hoc_diffractometer/geometry.py"
+      - "src/ad_hoc_diffractometer/diffractometer.py"
       - "src/ad_hoc_diffractometer/stage.py"
       - "src/ad_hoc_diffractometer/axes.py"
       - "src/ad_hoc_diffractometer/constants.py"
@@ -29,7 +29,7 @@ on:
       - "src/ad_hoc_diffractometer/reference.py"
       - "src/ad_hoc_diffractometer/presets.py"
       - "src/ad_hoc_diffractometer/factories.py"
-      - "src/ad_hoc_diffractometer/geometry.py"
+      - "src/ad_hoc_diffractometer/diffractometer.py"
       - "src/ad_hoc_diffractometer/stage.py"
       - "src/ad_hoc_diffractometer/axes.py"
       - "src/ad_hoc_diffractometer/constants.py"
