+++ docs/source/_static/custom.css
@@ -1,6 +1,6 @@
 /* In AutoAPI module pages, the "Module Contents" section shows inline
    class/function signatures with their fully-qualified module prefix
-   (sig-prename), e.g. "ad_hoc_diffractometer.geometry." before
+   (sig-prename), e.g. "ad_hoc_diffractometer.diffractometer." before
    "AdHocDiffractometer".  The full import path is already shown in the
    Import: container on each object's own page, so the prefix is
    redundant here.
