+++ docs/source/geometries/kappa4cv.md
@@ -156,7 +156,7 @@ before calling ``forward()``.
 ## API reference
 
 - {func}`~ad_hoc_diffractometer.presets.kappa4cv`
-- {class}`~ad_hoc_diffractometer.geometry.AdHocDiffractometer`
+- {class}`~ad_hoc_diffractometer.diffractometer.AdHocDiffractometer`
 - {class}`~ad_hoc_diffractometer.mode.ConstraintSet`
 - {class}`~ad_hoc_diffractometer.mode.BisectConstraint`
 - {class}`~ad_hoc_diffractometer.mode.SampleConstraint`
