+++ docs/source/geometries/s2d2.md
@@ -94,7 +94,7 @@ Requires ``g.surface_normal = (h, k, l)`` — see {doc}`../howto/surface`.
 ## API reference
 
 - {func}`~ad_hoc_diffractometer.presets.s2d2`
-- {class}`~ad_hoc_diffractometer.geometry.AdHocDiffractometer`
+- {class}`~ad_hoc_diffractometer.diffractometer.AdHocDiffractometer`
 - {class}`~ad_hoc_diffractometer.mode.ConstraintSet`
 - {class}`~ad_hoc_diffractometer.mode.SampleConstraint`
 - {class}`~ad_hoc_diffractometer.mode.ReferenceConstraint`
