+++ docs/source/quick_start.md
@@ -106,7 +106,7 @@ modes = {
 ## 4. Assemble the geometry
 
 Pass the stage list, basis, and modes to
-{class}`~ad_hoc_diffractometer.geometry.AdHocDiffractometer`:
+{class}`~ad_hoc_diffractometer.diffractometer.AdHocDiffractometer`:
 
 ```python
 g = ahd.AdHocDiffractometer(
@@ -196,7 +196,7 @@ See {doc}`geometries/fourcv` for the full geometry reference, or
 - {doc}`howto/custom_geometry` — how to build a geometry that is not one
   of the presets
 - {doc}`howto/basis_vectors` — choosing and understanding basis vectors
-- {class}`~ad_hoc_diffractometer.geometry.AdHocDiffractometer`
+- {class}`~ad_hoc_diffractometer.diffractometer.AdHocDiffractometer`
 - {class}`~ad_hoc_diffractometer.stage.Stage`
 - {class}`~ad_hoc_diffractometer.mode.ConstraintSet`
 - {class}`~ad_hoc_diffractometer.mode.BisectConstraint`
