+++ tests/test_mode.py
@@ -1219,7 +1219,7 @@ def test_mode_dict_values():
 
 
 # ---------------------------------------------------------------------------
-# AdHocDiffractometer — geometry.py lines 126, 132-133, 627-628, 647, 1857
+# AdHocDiffractometer — diffractometer.py lines 126, 132-133, 627-628, 647, 1857
 # ---------------------------------------------------------------------------
